+++ internal/provider/data_source_sdwan_application_priority_qos_policy_test.go
@@ -86,3 +86,6 @@ func testAccDataSourceSdwanApplicationPriorityQoSProfileParcelConfig() string {
 }
 
 // End of section. //template:end testAccDataSourceConfig
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSourceByNameConfig
+// End of section. //template:end testAccDataSourceByNameConfig