+++ internal/provider/data_source_sdwan_policy_object_security_identity_list_test.go
@@ -42,6 +42,14 @@ func TestAccDataSourceSdwanPolicyObjectSecurityIdentityListProfileParcel(t *test
 				Config: testAccDataSourceSdwanPolicyObjectSecurityIdentityListPrerequisitesProfileParcelConfig + testAccDataSourceSdwanPolicyObjectSecurityIdentityListProfileParcelConfig(),
 				Check:  resource.ComposeTestCheckFunc(checks...),
 			},
+			{
+				Config: testAccDataSourceSdwanPolicyObjectSecurityIdentityListPrerequisitesProfileParcelConfig + testAccDataSourceSdwanPolicyObjectSecurityIdentityListProfileParcelByNameConfig(),
+				Check: resource.ComposeTestCheckFunc(
+					append(checks,
+						resource.TestCheckResourceAttr("data.sdwan_policy_object_security_identity_list.test", "name", "TF_TEST"),
+						resource.TestCheckResourceAttrSet("data.sdwan_policy_object_security_identity_list.test", "id"),
+					)...),
+			},
 		},
 	})
 }
@@ -79,3 +87,25 @@ func testAccDataSourceSdwanPolicyObjectSecurityIdentityListProfileParcelConfig()
 }
 
 // End of section. //template:end testAccDataSourceConfig
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSourceByNameConfig
+func testAccDataSourceSdwanPolicyObjectSecurityIdentityListProfileParcelByNameConfig() string {
+	config := `resource "sdwan_policy_object_security_identity_list" "test" {` + "\n"
+	config += ` name = "TF_TEST"` + "\n"
+	config += ` description = "Terraform integration test"` + "\n"
+	config += `	feature_profile_id = sdwan_policy_object_feature_profile.test.id` + "\n"
+	config += `	entries = [{` + "\n"
+	config += `	  user = "administrator"` + "\n"
+	config += `	}]` + "\n"
+	config += `}` + "\n"
+
+	config += `
+		data "sdwan_policy_object_security_identity_list" "test" {
+			name = "TF_TEST"
+			feature_profile_id = sdwan_policy_object_feature_profile.test.id
+		}
+	`
+	return config
+}
+
+// End of section. //template:end testAccDataSourceByNameConfig