+++ internal/provider/data_source_sdwan_policy_object_security_scalable_group_tag_list_test.go
@@ -43,6 +43,14 @@ func TestAccDataSourceSdwanPolicyObjectSecurityScalableGroupTagListProfileParcel
 				Config: testAccDataSourceSdwanPolicyObjectSecurityScalableGroupTagListPrerequisitesProfileParcelConfig + testAccDataSourceSdwanPolicyObjectSecurityScalableGroupTagListProfileParcelConfig(),
 				Check:  resource.ComposeTestCheckFunc(checks...),
 			},
+			{
+				Config: testAccDataSourceSdwanPolicyObjectSecurityScalableGroupTagListPrerequisitesProfileParcelConfig + testAccDataSourceSdwanPolicyObjectSecurityScalableGroupTagListProfileParcelByNameConfig(),
+				Check: resource.ComposeTestCheckFunc(
+					append(checks,
+						resource.TestCheckResourceAttr("data.sdwan_policy_object_security_scalable_group_tag_list.test", "name", "TF_TEST"),
+						resource.TestCheckResourceAttrSet("data.sdwan_policy_object_security_scalable_group_tag_list.test", "id"),
+					)...),
+			},
 		},
 	})
 }
@@ -81,3 +89,26 @@ func testAccDataSourceSdwanPolicyObjectSecurityScalableGroupTagListProfileParcel
 }
 
 // End of section. //template:end testAccDataSourceConfig
+
+// Section below is generated&owned by "gen/generator.go". //template:begin testAccDataSourceByNameConfig
+func testAccDataSourceSdwanPolicyObjectSecurityScalableGroupTagListProfileParcelByNameConfig() string {
+	config := `resource "sdwan_policy_object_security_scalable_group_tag_list" "test" {` + "\n"
+	config += ` name = "TF_TEST"` + "\n"
+	config += ` description = "Terraform integration test"` + "\n"
+	config += `	feature_profile_id = sdwan_policy_object_feature_profile.test.id` + "\n"
+	config += `	entries = [{` + "\n"
+	config += `	  sgt_name = "ANY"` + "\n"
+	config += `	  tag = "65535"` + "\n"
+	config += `	}]` + "\n"
+	config += `}` + "\n"
+
+	config += `
+		data "sdwan_policy_object_security_scalable_group_tag_list" "test" {
+			name = "TF_TEST"
+			feature_profile_id = sdwan_policy_object_feature_profile.test.id
+		}
+	`
+	return config
+}
+
+// End of section. //template:end testAccDataSourceByNameConfig