+++ internal/provider/data_source_sdwan_system_ipv4_device_access_feature.go
@@ -24,11 +24,14 @@ import (
 	"net/url"
 
 	"github.com/CiscoDevNet/terraform-provider-sdwan/internal/provider/helpers"
+	"github.com/hashicorp/terraform-plugin-framework-validators/datasourcevalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/netascode/go-sdwan"
+	"github.com/tidwall/gjson"
 )
 
 // End of section. //template:end imports
@@ -37,8 +40,9 @@ import (
 
 // Ensure the implementation satisfies the expected interfaces.
 var (
-	_ datasource.DataSource              = &SystemIPv4DeviceAccessProfileParcelDataSource{}
-	_ datasource.DataSourceWithConfigure = &SystemIPv4DeviceAccessProfileParcelDataSource{}
+	_ datasource.DataSource                     = &SystemIPv4DeviceAccessProfileParcelDataSource{}
+	_ datasource.DataSourceWithConfigure        = &SystemIPv4DeviceAccessProfileParcelDataSource{}
+	_ datasource.DataSourceWithConfigValidators = &SystemIPv4DeviceAccessProfileParcelDataSource{}
 )
 
 func NewSystemIPv4DeviceAccessProfileParcelDataSource() datasource.DataSource {
@@ -61,7 +65,8 @@ func (d *SystemIPv4DeviceAccessProfileParcelDataSource) Schema(ctx context.Conte
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				MarkdownDescription: "The id of the Feature",
-				Required:            true,
+				Optional:            true,
+				Computed:            true,
 			},
 			"version": schema.Int64Attribute{
 				MarkdownDescription: "The version of the Feature",
@@ -69,6 +74,7 @@ func (d *SystemIPv4DeviceAccessProfileParcelDataSource) Schema(ctx context.Conte
 			},
 			"name": schema.StringAttribute{
 				MarkdownDescription: "The name of the Feature",
+				Optional:            true,
 				Computed:            true,
 			},
 			"description": schema.StringAttribute{
@@ -142,6 +148,15 @@ func (d *SystemIPv4DeviceAccessProfileParcelDataSource) Schema(ctx context.Conte
 	}
 }
 
+func (d *SystemIPv4DeviceAccessProfileParcelDataSource) ConfigValidators(_ context.Context) []datasource.ConfigValidator {
+	return []datasource.ConfigValidator{
+		datasourcevalidator.ExactlyOneOf(
+			path.MatchRoot("id"),
+			path.MatchRoot("name"),
+		),
+	}
+}
+
 func (d *SystemIPv4DeviceAccessProfileParcelDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
 	if req.ProviderData == nil {
 		return
@@ -164,6 +179,27 @@ func (d *SystemIPv4DeviceAccessProfileParcelDataSource) Read(ctx context.Context
 	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", config.Id.String()))
+	if config.Id.IsNull() && !config.Name.IsNull() {
+		// Look up parcel ID by name
+		res, err := d.client.Get(config.getPath())
+		if err != nil {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve parcels, got error: %s", err))
+			return
+		}
+		found := false
+		res.Get("data").ForEach(func(_, v gjson.Result) bool {
+			if v.Get("payload.name").String() == config.Name.ValueString() {
+				config.Id = types.StringValue(v.Get("parcelId").String())
+				found = true
+				return false
+			}
+			return true
+		})
+		if !found {
+			resp.Diagnostics.AddError("Not Found", fmt.Sprintf("No parcel found with name: %s", config.Name.ValueString()))
+			return
+		}
+	}
 
 	res, err := d.client.Get(config.getPath() + "/" + url.QueryEscape(config.Id.ValueString()))
 	if err != nil {