+++ internal/provider/provider.go
@@ -756,6 +756,7 @@ func (p *SdwanProvider) DataSources(ctx context.Context) []func() datasource.Dat
 		NewPolicerPolicyObjectDataSource,
 		NewPolicyGroupDataSource,
 		NewPolicyObjectFeatureProfileDataSource,
+		NewPolicyObjectFeatureProfileParcelsDataSource,
 		NewPortListPolicyObjectDataSource,
 		NewPreferredColorGroupPolicyObjectDataSource,
 		NewProtocolListPolicyObjectDataSource,